requirements.txt
lmdb==0.97
numpy==1.16.0
Pillow==6.1.0
six==1.12.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.59
opencv-python==3.4.0.12
Shapely==1.7.1
matplotlib==3.0.2
tensorflow-gpu==1.13.1 or tensorflow==1.13.1
Cython==0.29.6
ipython==5.1.0
scikit-image==0.17.2 
imutils==0.5.4
vs2015
pandas
pyqt5==5.15.6
openpyxl